refactor: add function to create the store
Help to recreate the store inside the tests.
