+++ src/index.js
@@ -1,12 +1,12 @@
+import { ConnectedRouter } from 'connected-react-router';
+import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import React from 'react';
-import { store, history } from './store';
-
 import { Route, Switch } from 'react-router-dom';
-import { ConnectedRouter } from 'connected-react-router';
 
 import App from './components/App';
+import history from './app/history';
+import store from './app/store';
 
 // https://www.cypress.io/blog/2018/11/14/testing-redux-store/
 /* istanbul ignore else */
