--- src/store.js
@@ -1,21 +1,0 @@
-import { configureStore } from '@reduxjs/toolkit';
-import { localStorageMiddleware } from './middleware';
-import createRootReducer from './reducer';
-
-import { routerMiddleware } from 'connected-react-router';
-import { createBrowserHistory } from 'history';
-
-export const history = createBrowserHistory();
-
-// Build the middleware for intercepting and dispatching navigation actions
-const myRouterMiddleware = routerMiddleware(history);
-
-export const store = configureStore({
-  reducer: createRootReducer(history),
-  devTools: process.env.NODE_ENV !== 'production',
-  middleware: getDefaultMiddleware => [
-    ...getDefaultMiddleware(),
-    myRouterMiddleware,
-    localStorageMiddleware,
-  ],
-});
